[Finished implementing the add new menu button]
